Update some controller links and remove stale links
Scenario guides no longer exist. They were removed in devel and will disappear from latest when Ansible 10 releases so changed a couple of links to point to the collection docs instead.

Also noticed a couple of controller links that should point to the controller docs on redhat.com now.

Lastly - there are a batch of links to Ansible.com at the bottom that won't exist when this website goes live. I haven't changed them here because I dunno if they have a new home somewhere on redhat.com or not.